Fix: Install missing DNS certbot plugins
While trying to setup the addon using DNSMadeEasy as a Let's Encrypt DNS
provider an error was run into. Looking through the open issues at the
time it was found that several issues had been raised with similar
problems.

After getting into the container while it was running it was discovered
that the only DNS provider installed was for Cloudflare. To verify that
the problem with DNSMadeEasy really was a missing plugin, it was
installed and then a new certificate was attempted and it succeeded.

This change brings certbot up to the latest upstream apk and installs
all of the DNS plugins that are named in the SSL Certificates system as
options.

Fixes: #251
Fixes: #243
Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>